requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
behave==1.2.6
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.1.3
execnet==2.0.2
fasteners==0.18
filelock==3.12.4
h11==0.14.0
idna==3.4
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parameterized==0.9.0
parse==1.19.1
parse-type==0.6.2
pdbp==1.5.0
platformdirs==3.10.0
pluggy==1.3.0
py==1.11.0
Pygments==2.16.1
pynose==1.4.8
pyotp==2.9.0
PySocks==1.7.1
pytest==7.4.2
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rich==13.5.2
sbvirtualdisplay==1.2.0
selenium==4.11.2
seleniumbase==4.18.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.0
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
xdist==0.0.2
